+++ doc/user/content/release-notes.md
@@ -61,6 +61,8 @@ Wrap your release notes at the 80 character mark.
 
 - Improve performance of `TopK` operator {{% gh 3758 %}}.
 
+- Handle Snappy-encoded Avro OCF files.
+
 <span id="v0.4.0"></span>
 ## v0.4.0